backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
torch
torchvision
pillow
python-multipart

